+++ asab/web/__init__.py
@@ -1,22 +1,19 @@
-import asab
-
-from .service import WebService
+from ..abc import Module
 from .websocket import WebSocketFactory
 from .staticdir import StaticDirProvider
-from .container import WebContainer
 
 
-class Module(asab.Module):
+class Module(Module):
 
 	def __init__(self, app):
 		super().__init__(app)
+
+		from .service import WebService
 		self.service = WebService(app, "asab.WebService")
 
 
 __all__ = (
-	'WebService',
 	'WebSocketFactory',
 	'StaticDirProvider',
-	'WebContainer',
 	'Module',
 )
