+++ asab/zookeeper/__init__.py
@@ -1,15 +1,15 @@
-from .service import ZooKeeperService
 from ..abc.module import Module
-from .container import ZooKeeperContainer
-from .builder import KazooWrapper
-__all__ = [
-	"ZooKeeperContainer",
-	"KazooWrapper",
-]
 
 
 class Module(Module):
 
 	def __init__(self, app):
 		super().__init__(app)
+
+		from .service import ZooKeeperService
 		self.Service = ZooKeeperService(app, "asab.ZooKeeperService")
+
+
+__all__ = [
+	"Module",
+]
