+++ setup.py
@@ -65,7 +65,6 @@ setup(
 	install_requires=[
 		'aiohttp>=3.6.2',
 		'fastjsonschema>=2.14.4',
-		'aiozk>=0.25.0',
 		'kazoo>=2.8.0',
 	],
 	cmdclass={
